Routes: Move Internship page to Past Programs
TEST=manual
