+++ js/src/patina/pages/past-programs/internship/FractalBadge.tsx
